Introduced proper Networkinit, Peer registration and User login, minor changes to db
db changes:
removed salt column

Networkkey and Id is now in a file managed by docker

+++ decentralizedCloud/Model/Entities/Data.cs
@@ -18,7 +18,7 @@ public class Data
     
     [Column("SIZE")]
     [Required]
-    public int Size { get; set; }//Byte
+    public long Size { get; set; }//Byte
     
     [Column("UPLOAD_TIME"),Required] 
     public DateTimeOffset UploadTime { get; init; }=DateTimeOffset.UtcNow;
