+++ decentralizedCloud/WebAPI/DTOs/UploadFileRequestDto.cs
@@ -5,5 +5,5 @@ public class UploadFileRequestDto
     public string Username { get; set; }
     public string Password { get; set; }
     public string FileName { get; set; }
-    public int FileSize { get; set; } // File size in bytes
+    public long FileSize { get; set; } // File size in bytes
 }
